Simplified actions structure
