--- client/store/actions/error.ts
@@ -1,7 +1,0 @@
-import * as store from '../store'
-
-export function closeError() {
-  store.setState('remove-error', (state) => {
-    state.error = undefined
-  })
-}
