+++ client/store/actions/file.ts
@@ -4,8 +4,7 @@ import * as helpers from '@client/helpers'
 import { cloneDeep } from 'lodash'
 import invariant from 'tiny-invariant'
 import * as store from '../store'
-import { openDialog } from './dialog'
-import { emitEvent } from './event'
+import { emitEvent, openDialog } from './app'
 import { loadSource } from './source'
 import { closeTable, getIsTableUpdated, openTable, revertTable } from './table'
