+++ client/store/actions/index.ts
@@ -1,10 +1,5 @@
 export * from './app'
-export * from './dialog'
-export * from './error'
-export * from './event'
 export * from './file'
 export * from './folder'
-export * from './panel'
-export * from './refs'
 export * from './resource'
 export * from './table'
