+++ client/store/index.ts
@@ -1,4 +1,3 @@
 export * from './actions'
-export { getRefs } from './actions/refs'
 export * from './state'
 export { getState, useStore } from './store'
